requirements.txt
aiomysql==0.2.0
aiosqlite==0.21.0
anyio==4.9.0
attrs==25.3.0
Automat==25.4.16
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
defusedxml==0.7.1
dnspython==2.7.0
fake-useragent==2.1.0
filelock==3.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iso8601==2.1.0
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
loguru==0.7.3
lxml==5.4.0
motor==3.7.0
packaging==25.0
parsel==1.10.0
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
pymongo==4.12.0
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pypika-tortoise==0.5.0
pytz==2025.2
PyYAML==6.0.2
queuelib==1.8.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
setuptools==80.0.0
sniffio==1.3.1
tldextract==5.3.0
tortoise-orm==0.24.2
tqdm==4.67.1
Twisted==24.11.0
typing_extensions==4.13.1
urllib3==2.4.0
w3lib==2.3.1
zope.interface==7.2
